+++ lib/src/annotation/parameter.dart
@@ -1,6 +1,7 @@
 import '../service_provider.dart';
 
 /// Provide an alternative parameter binding
+@Deprecated('Use @Inject(parameter: "name") instead.')
 class Parameter {
   /// The [name] of the bound parameter inside [ServiceProvider.parameters].
   final String name;