+++ example/lib/src/chat_provider.dart
@@ -1,3 +1,5 @@
+part of '../example.dart';
+
 abstract class ChatProvider {
   abstract final String username;