+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 Changes:
 - Updated the minimum `analyzer` version to `^5.0.0`
+
+Internal:
 - Updated the usages of deprecated properties. Using
   - `element` instead of `element2`
   - `enclosingElement` instead of `enclosingElement3`
+- Preflight logic refactored
 
 ## 3.1.0